+++ apps/api/src/modules/transactions/transactions.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { DatabaseModule } from '../../common/database/database.module';
 import { PromoModule } from '../promo/promo.module';
+import { CustomersModule } from '../customers/customers.module';
 import { TransactionsController } from './transactions.controller';
 import { TransactionsService } from './transactions.service';
 import { QrisPaymentService } from './qris-payment.service';
 
 @Module({
-  imports: [DatabaseModule, PromoModule],
+  imports: [DatabaseModule, PromoModule, CustomersModule],
   controllers: [TransactionsController],
   providers: [TransactionsService, QrisPaymentService],
   exports: [TransactionsService, QrisPaymentService],